requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
psycopg2-binary>=2.9.9
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0